+++ f1_server/CHANGELOG.md
@@ -1,8 +1,16 @@
 ### 1.1.0
 
+__Server__
+
 _Feat_
 
 - Init postgres databases and create function
+
+__Front__
+
+_Fix_
+
+- Change return of ranking Endpoint because object is not returned in this serverpod version
 
 # 1.0.0